+++ Packages/com.naukri.inspector-maid/Editor/Core/InspectorMaidEditor.cs
@@ -1,4 +1,5 @@
 ﻿using Naukri.InspectorMaid.Core;
+using Naukri.InspectorMaid.Editor.Events;
 using Naukri.InspectorMaid.Editor.Helpers;
 using Naukri.InspectorMaid.Editor.UIElements;
 using System.Collections.Generic;
@@ -14,17 +15,22 @@ namespace Naukri.InspectorMaid.Editor.Core
     {
         private VisualElement componentContainer;
 
+        public void RR()
+        {
+            componentContainer.SendEvent(new ChangeEvent<int>());
+        }
+
         public override VisualElement CreateInspectorGUI()
         {
-            componentContainer = new VisualElement() { name = "component" };
-            var styleSheets = InspectorMaidSettings.Instance.injectList;
+            componentContainer = new VisualElement() { name = "Component" };
+            var styleSheets = InspectorMaidSettings.Instance.importStyleSheets;
             foreach (var sheet in styleSheets)
             {
                 componentContainer.styleSheets.Add(sheet);
             }
 
             // fields
-            var fieldsContainer = new VisualElement() { name = "fields" };
+            var fieldsContainer = new VisualElement() { name = "Fields" };
             var iterator = serializedObject.GetIterator();
             if (iterator.NextVisible(true))
             {
@@ -40,7 +46,7 @@ namespace Naukri.InspectorMaid.Editor.Core
                     }
                     var fieldInfo = iterator.GetFieldInfo();
 
-                    if (fieldInfo.HasAttribute<InspectorMaidAttribute>())
+                    if (fieldInfo.HasAttribute<DecoratorAttribute>())
                     {
                         var fieldDrawer = new CustomFieldDrawer(target, fieldInfo, propertyField);
                         var fieldGUI = fieldDrawer.CreateFieldGUI();
@@ -57,12 +63,12 @@ namespace Naukri.InspectorMaid.Editor.Core
             var type = target.GetType();
 
             // properties
-            var propertiesContainer = new VisualElement() { name = "properties" };
+            var propertiesContainer = new VisualElement() { name = "Properties" };
             var propertyInfos = type.GetProperties(Utility.AllAccessFlags);
 
             foreach (var propertyInfo in propertyInfos)
             {
-                if (propertyInfo.HasAttribute<InspectorMaidAttribute>())
+                if (propertyInfo.HasAttribute<DecoratorAttribute>())
                 {
                     var propertyDrawer = new CustomPropertyDrawer(target, propertyInfo);
                     var propertyGUI = propertyDrawer.CreatePropertyGUI();
@@ -71,12 +77,12 @@ namespace Naukri.InspectorMaid.Editor.Core
             }
 
             // methods
-            var methodsContainer = new VisualElement() { name = "methods" };
+            var methodsContainer = new VisualElement() { name = "Methods" };
             var methodInfos = type.GetMethods(Utility.AllAccessFlags);
 
             foreach (var methodInfo in methodInfos)
             {
-                if (methodInfo.HasAttribute<InspectorMaidAttribute>())
+                if (methodInfo.HasAttribute<DecoratorAttribute>())
                 {
                     var methodDrawer = new CustomMethodDrawer(target, methodInfo);
                     var methodGUI = methodDrawer.CreatePropertyGUI();
@@ -97,28 +103,34 @@ namespace Naukri.InspectorMaid.Editor.Core
             {
                 foreach (var decorator in decorators)
                 {
-                    DecoratorElement.InvokeOnDestroy(decorator);
+                    Decorator.InvokeOnDestroy(decorator);
                 }
             }
         }
 
         protected void OnSceneGUI()
         {
-            if (TryGetDecoratorElements(out var decorators))
+            var needRepaint = true;
+            while (needRepaint)
             {
-                foreach (var decorator in decorators)
+                needRepaint = false;
+                if (TryGetDecoratorElements(out var decorators))
                 {
-                    if (!decorator.IsStarted)
+                    foreach (var decorator in decorators)
                     {
-                        DecoratorElement.InvokeOnStart(decorator);
-                        decorator.IsStarted = true;
+                        if (!decorator.IsStarted)
+                        {
+                            decorator.RegisterCallback<RepaintEvent>(evt => OnSceneGUI());
+                            Decorator.InvokeOnStart(decorator);
+                            decorator.IsStarted = true;
+                        }
+                        Decorator.InvokeOnSceneGUI(decorator);
                     }
-                    DecoratorElement.InvokeOnSceneGUI(decorator);
                 }
             }
         }
 
-        private bool TryGetDecoratorElements(out IEnumerable<DecoratorElement> decorators)
+        private bool TryGetDecoratorElements(out IEnumerable<Decorator> decorators)
         {
             if (componentContainer == null)
             {
@@ -127,7 +139,7 @@ namespace Naukri.InspectorMaid.Editor.Core
             }
             else
             {
-                decorators = componentContainer.Query<DecoratorElement>().ToList();
+                decorators = componentContainer.Query<Decorator>().ToList();
                 return true;
             }
         }