+++ Packages/com.naukri.inspector-maid/Editor/Events/InspectorMaidEvent.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine.UIElements;
+
+namespace Naukri.InspectorMaid.Editor.Events
+{
+    public class InspectorMaidEvent<T> : EventBase<T>
+        where T : EventBase<T>, new()
+    {
+    }
+}