+++ Packages/com.naukri.inspector-maid/Runtime/DividerAttribute.cs
@@ -1,0 +1,14 @@
+﻿using Naukri.InspectorMaid.Core;
+
+namespace Naukri.InspectorMaid
+{
+    public class DividerAttribute : ItemAttribute
+    {
+        public DividerAttribute(string text = null)
+        {
+            this.text = text;
+        }
+
+        public readonly string text;
+    }
+}