+++ Packages/com.naukri.inspector-maid/Samples/Attributes/Scripts/ContainerScopeSample.cs
@@ -1,12 +1,13 @@
 ﻿using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace Naukri.InspectorMaid.Samples
+namespace Naukri.InspectorMaid.Samples.Attributes
 {
     public class ContainerScopeSample : MonoBehaviour
     {
         [HelpBox("[ContainerScope] can help you to wrap your UI and add some custom styles.", HelpBoxMessageType.Info)]
         [ContainerScope, Style(margin: "10 0", padding: "5", backgroundColor: "#202020")]
+        // Sample 1
         [Target]
         [ContainerScope, Style(margin: "10 0", padding: "5", backgroundColor: "#404040", flexDirection: nameof(FlexDirection.Row))]
         [HelpBox("My UI1", HelpBoxMessageType.Info), Style(flexGrow: "0.5")]