+++ Packages/com.naukri.inspector-maid/Samples/Attributes/Scripts/TargetSample.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace Naukri.InspectorMaid.Samples
+namespace Naukri.InspectorMaid.Samples.Attributes
 {
     public class TargetSample : MonoBehaviour
     {
@@ -15,10 +15,9 @@ This is particularly useful when defining the target location or setting the tar
         public int withoutTarget = 0;
 
         [ContainerScope, Style(margin: "10 0", padding: "5", backgroundColor: "#202020")]
-        // Sample 1
+        // Sample 2
         [Target]
         [HelpBox("Due to the declaration of [Target] before [Helpbox], now the 'Helpbox' is positioned below 'Target'")]
         public int withTarget = 0;
     }
-
 }